# Single line pretty-printing options