--- a/internal/jsonflags/flags.go
+++ b/internal/jsonflags/flags.go
@@ -14,7 +14,7 @@ import "github.com/go-json-experiment/json/internal"
 //
 // In common usage, this is OR'd with 0 or 1. For example:
 //   - (AllowInvalidUTF8 | 0) means "AllowInvalidUTF8 is false"
-//   - (Expand | Indent | 1) means "Expand and Indent are true"
+//   - (Multiline | Indent | 1) means "Multiline and Indent are true"
 type Bools uint64
 
 func (Bools) JSONOptions(internal.NotForPublicUse) {}
@@ -67,7 +67,7 @@ const (
 		UnmarshalArrayFromAnyLength
 
 	// AnyWhitespace reports whether the encoded output might have any whitespace.
-	AnyWhitespace = Expand | SpaceAfterColon | SpaceAfterComma
+	AnyWhitespace = Multiline | SpaceAfterColon | SpaceAfterComma
 )
 
 // Encoder and decoder flags.
@@ -82,7 +82,7 @@ const (
 	CanonicalizeNumbers // encode only; for internal use by jsontext.Value.Canonicalize
 	EscapeForHTML       // encode only
 	EscapeForJS         // encode only
-	Expand              // encode only
+	Multiline           // encode only
 	SpaceAfterColon     // encode only
 	SpaceAfterComma     // encode only
 	Indent              // encode only; non-boolean flag