Discussion:
In the same vein as the `WithIndent` and `WithIndentPrefix` formatting options, I would really like the option to make json pretty but on a single line.
I end up reading, or at least skimming, through a lot of new-line delimited json, and this would make my life better 100%.

Consider reading the following:
```
{"foo": "bar", "baz": 123, "qux": {"quux": "corge"}, "grault": {"garply": ["waldo", "fred"]}, "thud": "plugh"}
{"foo":"bar", "baz":123, "qux":{"quux":"corge"}, "grault":{"garply":["waldo", "fred"]}, "thud":"plugh"}
{"foo":"bar","baz":123,"qux":{"quux":"corge"},"grault":{"garply":["waldo","fred"]},"thud":"plugh"}
```

The first two are much easier to see the key-value pairs and the elements than the third one, which is the only option right now.

https://github.com/golang/go/discussions/63397#discussioncomment-7300039


PR Details:
This PR adds two new boolean flag options: `SpaceAfterColon` and `SpaceAfterComma`
I do like those to be separate options, as I prefer to have the space after comma but not after colon, while lots of other people like space after both.
I'm not married to the naming or the implementation, and am open to any suggestions.
Please let me know if I need to add any additional tests beyond the two I added.