@dsnet 2023-11-02
Thanks for the PR.

I apologize for being a stickler for legal rules. Since the goal of this is intended to merge into the Go toolchain, we'll need every contribution to have also [signed the Go CLA](https://go.dev/doc/contribute#contributor). Is there a merged commit in any of the Go repositories that you can point to that proves you've signed the CLA? Thanks!

High-level thought: Modifying `MayAppendDelim` likely makes it such that it is no longer inlineable. We don't want to slow down the case where whitespace is completely elided. I'd like to see benchmarks of the performance delta of compacted output.