@veqryn 2023-11-03
Of course. I have signed the Go CLA. Here is my most recent contribution, though small:
https://github.com/golang/go/commit/65c53a1833a26467357b4aa6223e4dde5d6d7ed0
https://go-review.googlesource.com/c/go/+/524618