@veqryn 2023-11-03
I am also curious on how it may affect performance.  Do you want me to run benchmarks, and do those benchmarks already exist?  Or is this something you'd like to do to keep consistency with previous runs?